dom/nodes/CharacterData-appendData.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>CharacterData.appendData</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-appenddata">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-data">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testNode(create, type) {
  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.appendData("bar")
    assert_equals(node.data, "testbar")
  }, type + ".appendData('bar')")
}

testNode(function() { return document.createTextNode("test") }, "Text")
testNode(function() { return document.createComment("test") }, "Comment")
</script>
